fuzz: Add structured graph fuzz harness and fix folding overflows it found (#149)
* fuzz: Move build rules into a dedicated fuzz/Makefile

Keep the fuzzing tooling self contained so the production build never
pulls in clang or the sanitizer runtimes. The new fuzz/Makefile detects
the target, generates ir_fold_hash.h and the dynasm encoder header on
its own and builds every harness under fuzz/build.

Run the fuzz targets from the fuzz directory now.


* fuzz: Rename the text harness and standardize its corpus

The harness that drives the text parser was named fuzz_ir.c, which did
not say what set it apart since every harness here fuzzes the IR. Name
the source and its build targets after the input format. The source
becomes fuzz_text.c and the targets become fuzz-text-load and
fuzz-text-O0 through fuzz-text-O2, so the text harness reads apart from
the graph harness added next.

Keep a single seed corpus for the text targets instead of a directory
per optimization level. make_corpus.sh now writes the seeds to
corpus/text, which every text target shares, and corpus-clean only
removes those regenerable seeds. When no positional path is given on
the command line the harness falls back to corpus/text and creates it
on demand; an explicit path, such as a single crash file when
reproducing, is left untouched.


* fuzz: Add structured graph fuzz harness

This harness decodes the input blob directly into a structurally valid
IR function and runs the full optimization and codegen pipeline on it,
so inputs exercise the optimizer and backend.

The input bytes are turned into a valid IR function step by step. The
first byte picks the type and parameter count, then each small group
of bytes adds one operation that uses earlier nodes as inputs. The
last node becomes the return value.

The optimization level is selected at compile time via FUZZ_GRAPH_O0,
FUZZ_GRAPH_O1 or FUZZ_GRAPH_O2.


* fuzz: Document the standalone build and the graph harness

Point every command at the fuzz directory now that the build lives in
fuzz/Makefile and add the fuzz-graph targets to the quick start, the
target table and the crash reproduction notes.


* ir: Avoid signed overflow in MUL(C_U16, C_U16) folding

The two u16 operands are promoted to signed int before the multiply,
so a wide product such as 54300 * 54300 overflows int. The result is
truncated back to u16 and is correct on any two complement target,
but it is still undefined behavior.

Cast both operands to uint32_t so the multiply is unsigned.


* ir: Avoid signed overflow in MUL/DIV(NEG, C_I) folding

The rules MUL(NEG(x), c) and DIV(NEG(x), c) negate the constant on a
signed int64_t, which is undefined behavior for INT64_MIN. Negate
through uint64_t instead. The result is unchanged on a release build
since the negation of INT64_MIN wraps to itself.


---------

+++ ir_fold.h
@@ -524,7 +524,7 @@ IR_FOLD(MUL(C_U8, C_U8))
 IR_FOLD(MUL(C_U16, C_U16))
 {
 	IR_ASSERT(IR_OPT_TYPE(opt) == op1_insn->type);
-	IR_FOLD_CONST_U(op1_insn->val.u16 * op2_insn->val.u16);
+	IR_FOLD_CONST_U((uint32_t)op1_insn->val.u16 * (uint32_t)op2_insn->val.u16);
 }
 
 IR_FOLD(MUL(C_U32, C_U32))
@@ -2216,7 +2216,7 @@ IR_FOLD(DIV(NEG, C_I32))
 IR_FOLD(DIV(NEG, C_I64))
 {
 	op1 = op1_insn->op1;
-	val.i64 = -op2_insn->val.i64;
+	val.i64 = -(uint64_t)op2_insn->val.i64;
 	op2 = ir_const(ctx, val, op2_insn->type);
 	IR_FOLD_RESTART;
 }
